Refactor: extract DropdownMenu context.ts and utils.ts
